codigo do notebook modularizado
